assignReviews ought to be working
It makes a blank review for each student/submission pair. This review
row will not have any filename, or contents. Hopefully everyone will
just ignore it.

The idea is that if a review exists for a student/submission pair, then
they must have to review that submission. More review rows may be made
at any time, by the student marking actual files within the submission,
but this will not affect anything. Additionally, they will be able to
save the reviews, and come back to them later, or submit them for the
other student to see, so there's no issue.

Of course, a student will never be assigned their own submission to
mark.